+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/common/classId.kt
@@ -1,4 +1,4 @@
-package com.javiersc.kotlin.compiler.extensions.ir
+package com.javiersc.kotlin.compiler.extensions.common
 
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
@@ -7,5 +7,5 @@ public fun String.toClassId(): ClassId = FqName(this).run { ClassId(parent(), sh
 
 public fun FqName.toClassId(): ClassId = ClassId(parent(), shortName())
 
-public inline fun <reified T> classId(): ClassId = com.javiersc.kotlin.compiler.extensions.common.fqName<T>()
+public inline fun <reified T> classId(): ClassId = fqName<T>()
     .toClassId()