+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irElement.kt
@@ -1,6 +1,7 @@
 package com.javiersc.kotlin.compiler.extensions.ir
 
 import org.jetbrains.kotlin.ir.IrElement
+import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
 import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
 import org.jetbrains.kotlin.ir.util.dump
@@ -12,9 +13,13 @@ public fun Iterable<IrElement>.dump(): String = joinToString("\n") { it.dump() }
 
 public fun Iterable<IrElement>.dumpKotlinLike(): String = joinToString("\n") { it.dumpKotlinLike() }
 
+public inline fun <reified T : Annotation> IrElement.hasAnnotation(): Boolean =
+    hasAnnotation(T::class.toFqName())
+
 public fun IrElement.hasAnnotation(annotation: FqName): Boolean =
     when (this) {
         is IrClass -> hasAnnotation(annotation)
         is IrFunctionAccessExpression -> hasAnnotation(annotation)
+        is IrAnnotationContainer -> hasAnnotation(annotation)
         else -> false
     }