+++ kotlin-compiler-extensions/test-data/diagnostics/playground-test.kt
@@ -1,6 +1,0 @@
-// FIR_IDENTICAL
-package foo.bar
-
-fun greetings(): String {
-    return "hello"
-}